requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==41.0.3
cycler==0.10.0
decorator==4.4.0
EditorConfig==0.12.3
entrypoints 
idna==2.8
imageio==2.5.0
ipykernel 
ipython 
ipython-genutils 
jedi==0.15.0
joblib==0.13.2
jsbeautifier==1.14.9
jupyter-client==5.3.4
jupyter-core==4.5.0
kiwisolver==1.1.0
matplotlib==3.1.0
matplotlib-inline
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
olefile==0.46
opencv-python==4.8.0.76
pandas==0.25.0
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare
Pillow==9.5.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyarrow==12.0.1
pycparser==2.19
Pygments 
pyOpenSSL==23.2.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.2
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.1
torch==1.13.0
torchfile==0.1.0
torchvision==0.14.0
tornado==5.1.1
tqdm==4.33.0
traitlets==4.3.2
typing_extensions==4.7.1
urllib3==1.25.11
visdom==0.1.8.8
wcwidth==0.1.7
websocket-client==0.56.0
xmltodict==0.13.0
